refactor: Player와 Dealer가 공통의 추상 클래스 Participant를 상속받도록 변경
