+++ src/main/java/blackjack/domain/player/Player.java
@@ -1,38 +1,12 @@
 package blackjack.domain.player;
 
-import blackjack.domain.card.Card;
-import java.util.List;
-
-public class Player {
-    protected final Hand hand;
-    private final Name name;
-
-    public Player(final String name) {
-        this.hand = new Hand();
-        this.name = new Name(name);
+public class Player extends Participant {
+    public Player(String name) {
+        super(name);
     }
 
-    public void addCard(final Card card) {
-        hand.add(card);
-    }
-
-    public boolean isBust() {
-        return hand.isBust();
-    }
-
-    public boolean isBlackjack() {
-        return hand.isBlackjack();
-    }
-
-    public int getScore() {
-        return hand.calculateScore();
-    }
-
-    public List<Card> getCards() {
-        return hand.getAllCards();
-    }
-
-    public String getName() {
-        return name.getValue();
+    @Override
+    public boolean isDrawable() {
+        return !isBust();
     }
 }
